refactor: 멤버 의존성 정리 프로필 조회 분리
