+++ backend/src/main/java/com/yigongil/backend/application/AuthService.java
@@ -3,8 +3,8 @@ package com.yigongil.backend.application;
 import com.yigongil.backend.config.auth.JwtTokenProvider;
 import com.yigongil.backend.config.oauth.GithubOauth;
 import com.yigongil.backend.config.oauth.GithubProfileResponse;
-import com.yigongil.backend.domain.member.Member;
-import com.yigongil.backend.domain.member.MemberRepository;
+import com.yigongil.backend.domain.member.domain.Member;
+import com.yigongil.backend.domain.member.domain.MemberRepository;
 import com.yigongil.backend.request.TokenRequest;
 import com.yigongil.backend.response.TokenResponse;
 import java.util.Optional;
