+++ backend/src/main/java/com/yigongil/backend/domain/member/domain/Introduction.java
@@ -1,4 +1,4 @@
-package com.yigongil.backend.domain.member;
+package com.yigongil.backend.domain.member.domain;
 
 import com.yigongil.backend.exception.InvalidIntroductionLengthException;
 import java.util.Objects;
