+++ backend/src/main/java/com/yigongil/backend/domain/member/ui/MemberController.java
@@ -1,18 +1,17 @@
-package com.yigongil.backend.domain.member;
+package com.yigongil.backend.domain.member.ui;
 
 import com.yigongil.backend.config.auth.Authorization;
+import com.yigongil.backend.domain.member.application.MemberService;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.request.ProfileUpdateRequest;
-import com.yigongil.backend.response.MyProfileResponse;
 import com.yigongil.backend.response.NicknameValidationResponse;
 import com.yigongil.backend.response.OnboardingCheckResponse;
-import com.yigongil.backend.response.ProfileResponse;
 import com.yigongil.backend.ui.doc.MemberApi;
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -26,30 +25,6 @@ public class MemberController implements MemberApi {
 
     public MemberController(MemberService memberService) {
         this.memberService = memberService;
-    }
-
-    @GetMapping(path = "/{id}")
-    public ResponseEntity<ProfileResponse> findProfile(@PathVariable Long id) {
-        ProfileResponse response = memberService.findById(id);
-        return ResponseEntity.ok(response);
-    }
-
-    @GetMapping(path = "/my")
-    public ResponseEntity<MyProfileResponse> findMyProfile(@Authorization Member member) {
-        ProfileResponse profile = memberService.findById(member.getId());
-        MyProfileResponse response = new MyProfileResponse(
-                profile.id(),
-                profile.nickname(),
-                profile.githubId(),
-                profile.profileImageUrl(),
-                profile.successRate(),
-                profile.successfulRoundCount(),
-                profile.tierProgress(),
-                profile.tier(),
-                profile.introduction()
-        );
-
-        return ResponseEntity.ok(response);
     }
 
     @PatchMapping
