+++ backend/src/main/java/com/yigongil/backend/domain/round/HomeController.java
@@ -1,7 +1,7 @@
-package com.yigongil.backend.domain.member;
+package com.yigongil.backend.domain.round;
 
 import com.yigongil.backend.config.auth.Authorization;
-import com.yigongil.backend.domain.round.RoundService;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.response.UpcomingStudyResponse;
 import com.yigongil.backend.ui.doc.HomeApi;
 import java.util.List;
