+++ backend/src/main/java/com/yigongil/backend/domain/round/MustDoController.java
@@ -1,7 +1,7 @@
 package com.yigongil.backend.domain.round;
 
 import com.yigongil.backend.config.auth.Authorization;
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.request.MustDoUpdateRequest;
 import com.yigongil.backend.ui.doc.MustDoApi;
 import javax.validation.Valid;
