+++ backend/src/main/java/com/yigongil/backend/domain/round/RoundService.java
@@ -1,6 +1,6 @@
 package com.yigongil.backend.domain.round;
 
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.study.ProcessingStatus;
 import com.yigongil.backend.domain.study.Study;
 import com.yigongil.backend.domain.study.StudyRepository;
