+++ backend/src/main/java/com/yigongil/backend/domain/study/Study.java
@@ -2,7 +2,7 @@ package com.yigongil.backend.domain.study;
 
 import com.yigongil.backend.domain.base.BaseEntity;
 import com.yigongil.backend.domain.meetingdayoftheweek.MeetingDayOfTheWeek;
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.roundofmember.RoundOfMember;
 import com.yigongil.backend.domain.studymember.Role;
