+++ backend/src/main/java/com/yigongil/backend/domain/study/StudyService.java
@@ -3,7 +3,7 @@ package com.yigongil.backend.domain.study;
 import com.yigongil.backend.domain.certification.Certification;
 import com.yigongil.backend.domain.certification.CertificationService;
 import com.yigongil.backend.domain.feedpost.FeedService;
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.round.RoundRepository;
 import com.yigongil.backend.domain.roundofmember.RoundOfMember;
