+++ backend/src/main/java/com/yigongil/backend/ui/doc/HomeApi.java
@@ -1,6 +1,6 @@
 package com.yigongil.backend.ui.doc;
 
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.response.UpcomingStudyResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
