+++ backend/src/main/java/com/yigongil/backend/ui/doc/ReportApi.java
@@ -1,7 +1,7 @@
 package com.yigongil.backend.ui.doc;
 
 import com.yigongil.backend.config.auth.Authorization;
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.request.MemberReportCreateRequest;
 import com.yigongil.backend.request.StudyReportCreateRequest;
 import io.swagger.v3.oas.annotations.Operation;
