+++ backend/src/test/java/com/yigongil/backend/fixture/StudyFixture.java
@@ -1,6 +1,6 @@
 package com.yigongil.backend.fixture;
 
-import com.yigongil.backend.domain.member.Member;
+import com.yigongil.backend.domain.member.domain.Member;
 import com.yigongil.backend.domain.round.Round;
 import com.yigongil.backend.domain.study.ProcessingStatus;
 import com.yigongil.backend.domain.study.Study;
